feat: flytt Sentry-init til SøknadAppProvider via sentryConfig-prop
- Legg til initSentry() og SentryConfig i sif-common-sentry/v2.ts
- Trekk ut isErrorFromDekoratøren og beforeSendFilter til sentryFilters.ts
  slik at v1 (index.ts) og v2 (v2.ts) ikke er avhengige av hverandre
- Legg til sentryConfig-prop i SøknadAppProvider med synkron init og
  modul-level guard mot dobbeltkall
- Fjern sentry/instrument.ts fra aktivitetspenger-soknad
- Oppdater initApiClients.ts til å importere setRedirectingToLogin
  direkte fra @navikt/sif-common-sentry
