+++ apps/aktivitetspenger-soknad/src/main.tsx
@@ -1,5 +1,3 @@
-import './sentry/instrument';
-
 import { reactErrorHandler } from '@sentry/react';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
