+++ packages/sif-common-sentry/src/index.ts
@@ -1,6 +1,8 @@
 import * as Sentry from '@sentry/browser';
 import { AxiosError } from 'axios';
 
+import { beforeSendFilter } from './sentryFilters';
+
 export enum SentryEnvironment {
     LOCALHOST = 'LOCALHOST',
     q = 'q',
@@ -118,34 +120,6 @@ export const setupIgnoreErrorsAndAllowUrls = (
         /* empty */
     }
     return { allowUrls, ignoreErrors };
-};
-
-const dekoratorenTimeoutPatterns = ['Request timeout', 'dekoratoren'];
-
-export const isErrorFromDekoratøren = (event: Sentry.ErrorEvent): boolean => {
-    const values = event.exception?.values ?? [];
-
-    const frames = values.flatMap((v) => v.stacktrace?.frames ?? []);
-    if (frames.some((f) => (f.filename ?? '').includes('/dekoratoren/'))) {
-        return true;
-    }
-
-    const firstValue = values[0];
-    if (firstValue?.type === 'UnhandledRejection') {
-        const message = firstValue.value ?? '';
-        if (dekoratorenTimeoutPatterns.some((pattern) => message.includes(pattern))) {
-            return true;
-        }
-    }
-
-    return false;
-};
-
-export const beforeSendFilter = (event: Sentry.ErrorEvent): Sentry.ErrorEvent | null => {
-    if (isErrorFromDekoratøren(event)) {
-        return null;
-    }
-    return event;
 };
 
 const initSentryForSIF = (initProps: SentryInitProps = {}) => {
