+++ packages/sif-soknad-app/package.json
@@ -22,6 +22,7 @@
         "@navikt/ds-react": "8.11.1",
         "@navikt/nav-dekoratoren-moduler": "4.2.1",
         "@navikt/sif-common-faro": "workspace:*",
+        "@navikt/sif-common-sentry": "workspace:*",
         "@navikt/sif-common-ui": "workspace:*",
         "@sentry/react": "10.56.0",
         "@sif/api": "workspace:*",
