+++ packages/sif-soknad-app/src/components/SøknadAppProvider.tsx
@@ -1,3 +1,4 @@
+import { initSentry, SentryConfig } from '@navikt/sif-common-sentry';
 import { FaroProvider } from '@navikt/sif-common-faro';
 import { DevBranchInfo } from '@sif/soknad-ui';
 import { PropsWithChildren } from 'react';
@@ -11,15 +12,24 @@ interface SøknadAppProviderProps {
     appVersion: string;
     isActive?: boolean;
     telemetryCollectorURL?: string;
+    sentryConfig?: SentryConfig;
 }
+
+let sentryInitialized = false;
 
 export const SøknadAppProvider = ({
     applicationKey,
     appVersion,
     isActive,
     telemetryCollectorURL,
+    sentryConfig,
     children,
 }: PropsWithChildren<SøknadAppProviderProps>) => {
+    if (sentryConfig && !sentryInitialized) {
+        initSentry(sentryConfig);
+        sentryInitialized = true;
+    }
+
     return (
         <FaroProvider
             applicationKey={applicationKey}
